requirements.txt
fastapi
uvicorn
pydantic>=2
openai
filelock
PyYAML